+++ changelog.md
@@ -1,3 +1,4 @@
 ## General
 
 - ***Added*** `/ad` as a shortcut for `/address`
+- Translations are now loaded on the server side, so no more fabric-api needed